airflows/requirements.txt
apache-airflow==2.4.3
psycopg2-binary==2.8.6
mysql-connector-python==8.0.27  # MySQL bağlantısı için
scikit-learn==0.24.2
matplotlib==3.3.3